Module Resto

module Utils : sig ... end

type meth = [

| `GET

| `POST

| `DELETE

| `PUT

| `PATCH

]

val string_of_meth : [< meth ] -> string

val meth_of_string : string -> [> meth ] option

module MethMap : Stdlib.Map.S with type key = meth

module StringMap : Stdlib.Map.S with type key = string

type (_, _) eq =

| Eq : ('a, 'a) eq

eq is an equality witness type. It is returned by some non-trivial equality-testing functions in the rest of this module. In general, it is intended to be used as follows:

match are_equal foo bar with | None -> (* values are not equal *) .. | Some Eq -> (* values are equal *) ..

module Arg : sig ... end

Typed path argument.

module Path : sig ... end

Parametrized path to services.

module Description : sig ... end

Service directory description

module Query : sig ... end

module type ENCODING = sig ... end

An ENCODING is a generic interface for modules that provide conversion from values to a different representation and back. This is used to abstract resto over specific representations of values.

module MakeService (Encoding : ENCODING) : sig ... end